% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.generator.R
\name{cbs}
\alias{cbs}
\title{Generate Constant B-spline Basis}
\usage{
cbs(x, knots, Boundary.knots = range(x))
}
\arguments{
\item{x}{A numeric vector for which the constant spline basis is to be generated.}

\item{knots}{A numeric vector of knots that define the intervals for the constant spline basis.}

\item{Boundary.knots}{A numeric vector of length two defining the boundary knots. Default is the range of `x`. These knots represent the outer boundaries for the spline basis.}
}
\value{
A matrix of binary indicators with rows corresponding to the input vector `x` and columns corresponding to the defined knots. Each entry is `1` if the corresponding value in `x` falls within the interval defined by the knots and `0` otherwise.
}
\description{
This function creates a constant spline basis matrix for a given vector of input values. It assigns binary indicators for which interval each value falls into based on specified knots.
}
\details{
The function ensures that there is no overlap between the defined intervals by assigning values only to the corresponding interval where each input falls.
The first column of the output matrix corresponds to values less than the first knot, and the last column corresponds to values greater than or equal to the last knot.
}
