% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Simulate Data for Classification Problems}
\usage{
simulation(n.class, case.num)
}
\arguments{
\item{n.class}{An integer indicating the number of classes for the simulation. Acceptable values are 2 or 3.}

\item{case.num}{An integer specifying the case number of the simulation to execute. This should correspond to the specific simulation scenarios available for the selected number of classes.}
}
\value{
A list containing:
  \item{train}{A data frame of the training dataset generated from the specified simulation case.}
  \item{test}{A data frame of the testing dataset generated from the specified simulation case.}
}
\description{
This function simulates datasets for classification tasks based on the specified number of classes and case number. It sources external simulation scripts to generate training and testing datasets.
}
\details{
The function will source R scripts from predetermined paths based on the number of classes (2 or 3).
For each class scenario, the function can simulate up to four different cases, each generating distinct datasets.
It is important that the simulation scripts exist in the specified directories, and they should return a list containing `train.df` and `test.df`.
There are four different cases for simulation:
 (1) the groups are distinct across all time points, exhibiting similar average trends;
 (2) the groups diverge only within a specific time window, from the time point [5, 15], displaying different patterns;
 (3) the groups show separation over a fixed interval of 10 time points, though the exact time frame varies;
 (4) the groups are distinct during a random occurring period, with the length of the separation varying between 5 and 40 time points.
}
\examples{
# Simulate data for a 2-class problem, case 1
result <- simulation(n.class = 2, case.num = 1)

# Access training and testing datasets
train_data <- result$train
test_data <- result$test

}
